+++ CHANGES.md
@@ -10,6 +10,9 @@
 
 ### Improvements
 
+-   Add `datetime_as_string` and `mixed_offsets_as_utc` parameters to `read_dataframe`
+    to choose the way datetime columns are returned + several fixes when reading and
+    writing datetimes (#486).
 -   Add listing of GDAL data types and subtypes to `read_info` (#556).
 -   Add support to read list fields without arrow (#558).